Add new EventManager module
The EventManager is an interface for managing plugin events. The common
issue with Bukkit event systems is that the events are quite difficult to
fire: you often need to synchronize the method calls and hold the Plugin
instance, so this abstraction offers a solition for this problem. Using
both synchronized and asynchronous methods from the EventManager
interface allows programmer to easily manage the event calls.